Rename q2_gradcheck.py to q2_gradcheck_old.py
This solution is not working. It won't work for q2_neuron.py as that requires all parameter as the inputs
check url below for explanation
https://www.reddit.com/r/CS224d/comments/3477wo/issue_with_gradcheck_naive_for_forward_backward/
''
yes, remember that x is multi-dimensional variable and f is a multivariate function. The idea of the gradient check is for each dimension, ix, we want to estimate the gradient in that dimension alone. Your previous code didn't work because x[ix] + h has dimension 1.
''